Continued refactoring page components
