+++ src/Routers/Router.js
@@ -5,9 +5,9 @@ import AblComponent from "../pages/abl/abl.component";
 import ArchitectComponent from "../pages/architect/architect.component";
 import Cobas8000Component from "../pages/cobas8000/cobas8000.component";
 import Cobas8100Component from "../pages/cobas8100/cobas8100.component";
-import ManualUsedTransaction from "../pages/ManualUsedTransaction";
-import Error404 from "../pages/404";
-import ItemDetails from "../pages/ItemDetails";
+import ManualUsedTransaction from "../pages/manual-used-transaction/manual-used-transaction.component";
+import Error404 from "../pages/not-found/not-found.component";
+import ItemDetails from "../pages/item-details/item-details.component";
 import SidebarComponent from "../components/sidebar/sidebar.component";
 import { ArchitectContext } from "../Contexts/ArchitectContext";
 import { ABLContext } from "../Contexts/ABLContext";
