+++ src/pages/item-details/item-details.styles.scss
